algorithm now finds the alignment pattern and uses it to transform and crop the image

+++ Sobel.m
@@ -8,7 +8,7 @@ assert(isa(originalImage, 'double'));
 assert(isa(threshold, 'double'));
 
 k = [1 2 1; 0 0 0; -1 -2 -1];
-H = conv2(double(originalImage),k, 'same');
-V = conv2(double(originalImage),k','same');
+H = conv2(originalImage,k, 'same');
+V = conv2(originalImage,k','same');
 E = sqrt(H.*H + V.*V);
 edgeImage = logical((E > threshold) * 255);
